add F329 workspace with new motorhand
using Shadow's melodic branches, but the machine still runs Ubuntu 16.04 (ROS kinetic),
so we need to revert/adapt some things.

I actually use this together with a MoveIt `master` workspace,
but hopefully this is not relevant for others.
Also you might have to revert some commits in MoveIt's master branches, because they broke compatibility with Ubuntu 16.04
end of August 2020.